🔵 other: Moving login components into app/login folder
